+++ kernel/src/gui/basics.c
@@ -18,16 +18,15 @@ void draw_rectangle(const uint32_t x, uint32_t y, uint32_t w, uint32_t h, uint32
 
 void draw_filled_rectangle(size_t x, size_t y, size_t w, size_t h, uint32_t fill) {
     for(size_t i = 0; i < h; i++) {
-        register int j = 0;
+        register size_t j = 0;
 
-        for(; j < w; j+=4) {
-            //set_pixel(x+j, y+i, fill);
-            set_pixel4x1(x+j, y+i, fill);
+        for(; j < w; j += 4) {
+            set_pixel4x1(x + j, y + i, fill);
         }
 
         j -= 4;
 
-        for(; j<w; j++) {
+        for(; j < w; j++) {
             set_pixel(x+j, y+i, fill);
         }
     }
